requirements.txt
flask==2.3.3
google-generativeai==0.3.2
playwright==1.40.0
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
beautifulsoup4==4.12.2
numpy==1.25.2
scipy==1.11.4
requests==2.31.0
python-dotenv==1.0.0
